Update README.md: grammar, spelling, style
Corrected some typos and such.
Changed "OpenHardwareMonitor" to "Open Hardware Monitor" because that is how the project calls itself on its GitHub and on its website.
Added spaces before a lot of parentheses. While I do realise that the lack of spacing was intended for context clarity, it doesn't really seem necessary? The non-spacing didn't seem necessary for formatting/linebreak reasons.